requirements.txt
fastapi>=0.75.0
uvicorn>=0.17.0
celery>=5.2.0
redis>=4.3.0
pillow>=9.0.0
python-multipart>=0.0.5
torch>=1.12.0
torchvision>=0.13.0
transformers
--extra-index-url https://download.pytorch.org/whl/cpu